+++ src/render/Fader.java
@@ -5,21 +5,21 @@ public class Fader extends DrawableImage {
 	
 	// Members 
 	private boolean m_fading = false;
-	private double[][] m_targets = {{1.0,100}};
+	private float[][] m_targets = {{1.0f,100f}};
 	private int m_tid = 0;
-	private double m_start = 1;
-	private long m_fx = 0, m_duration = 100;
+	private float m_start = 1f;
+	private int m_fx = 0, m_duration = 100;
 	
-	public Fader(int img, double x, double y, double w, double h) {
+	public Fader(int img, float x, float y, float w, float h) {
 		super(img, x, y, w, h);
 		
 	}
 	
-	public void update(long timePassed) {		
+	public void update(int timePassed) {		
 		if (m_fading) {
 			m_fx += timePassed;
 			
-			setAlpha(fadeFunction((double)m_fx,m_start,m_targets[m_tid][0]-m_start,(double)m_targets[m_tid][1]));
+			setAlpha(fadeFunction((float)m_fx,m_start,m_targets[m_tid][0]-m_start,(float)m_targets[m_tid][1]));
 			
 			if (m_fx >= m_duration) {
 				m_fading = false;
@@ -34,19 +34,19 @@ public class Fader extends DrawableImage {
 		}
 	}
 	
-	private double fadeFunction(double t, double b, double c, double d) {
-		return Math.min(Math.max((-c) * Math.cos((t/d) * (Math.PI/2)) + c + b, 0),1);
+	private float fadeFunction(float t, float b, float c, float d) {
+		return (float) Math.min(Math.max((-c) * Math.cos((t/d) * (Math.PI/2)) + c + b, 0),1);
 	}
 	
 	public boolean isFading() {
 		return m_fading;
 	}
 	
-	public void setFadeTarget(double t) {
-		setFadeTarget(new double[][]{{t,m_duration}});
+	public void setFadeTarget(float t) {
+		setFadeTarget(new float[][]{{t,m_duration}});
 	}
 	
-	public void setFadeTarget(double[][] t) {
+	public void setFadeTarget(float[][] t) {
 		m_fx = 0;
 		m_tid = 0;
 		m_targets = t;
